Split delete and common toolbar actions mapping
