+++ app/javascript/helpers/rxConnector.js
@@ -1,12 +1,15 @@
+import { listenToRx } from '../miq_observable';
+
 function callMapperFunction(eventMapper, event) {
   return Object.prototype.hasOwnProperty.call(eventMapper, event.type)
     && eventMapper[event.type](event.payload);
 }
 
 export function subscribeToRx(eventMapper) {
-  ManageIQ.angular
-    .rxSubject
-    .subscribe(event => event.type && callMapperFunction(eventMapper, event));
+  listenToRx(event => event.type && callMapperFunction(eventMapper, event));
 }
 
-export const DELETE_EVENT = 'delete';
+/**
+ * Event types are stored here.
+ */
+export const DELETE_EVENT = 'delete'; // Reacts to event - {type: 'delete', payload: {...}}
