+++ app/javascript/tests/toolbar-actions-delete.test.js
@@ -1,14 +1,31 @@
 import 'angular-mocks';
-import * as toolbarActions from '../packs/toolbar-actions';
+import * as commonActions from '../packs/toolbar-actions';
+import * as deleteActions from '../toolbar-actions/delete';
 
 describe('Toolbar actions', () => {
   beforeEach(() => {
     angular.mock.module('miq.api');
   });
 
+  describe('gridChecks', () => {
+    beforeEach(() => {
+      window.ManageIQ.gridChecks = [];
+    });
+
+    it('should use ManageIQ.gridChecks if any present', () => {
+      window.ManageIQ.gridChecks = ['something', '15586'];
+      expect(commonActions.getGridChecks()).toEqual([{ id: 'something' }, { id: '15586' }]);
+    });
+
+    it('should use ManageIQ.record if no gridChecks', () => {
+      window.ManageIQ.record = { recordId: 'something' };
+      expect(commonActions.getGridChecks()).toEqual([{ id: 'something' }]);
+    });
+  });
+
   describe('divideNames', () => {
     it('should count success and failures', () => {
-      const messageCount = toolbarActions.generateMessages([
+      const messageCount = deleteActions.generateMessages([
         { success: false, message: 'Some message' },
         { success: false, message: 'Some message' },
         { success: true, message: 'Some message' },
@@ -18,7 +35,7 @@ describe('Toolbar actions', () => {
     });
 
     it('should have empty items for wrong messages', () => {
-      const messageCount = toolbarActions.generateMessages([
+      const messageCount = deleteActions.generateMessages([
         { message: 'No success' },
       ]);
       expect(messageCount).toEqual({ 'false': 0, 'true': 0 });
@@ -35,14 +52,14 @@ describe('Toolbar actions', () => {
     });
 
     it('calls show message with string message', () => {
-      toolbarActions.showMessage('some message', false);
+      deleteActions.showMessage('some message', false);
       expect(addFlash).toHaveBeenCalledWith('some message', 'error');
-      toolbarActions.showMessage('some message 2', true);
+      deleteActions.showMessage('some message 2', true);
       expect(addFlash).toHaveBeenCalledWith('some message 2', 'success');
     });
 
     it('calls show message with object', () => {
-      toolbarActions.showMessage({ false: 1, true: 4 });
+      deleteActions.showMessage({ false: 1, true: 4 });
       expect(window.sprintf).toHaveBeenCalled();
       expect(addFlash).toHaveBeenCalledWith('Deleting of 4 items queued.', 'success');
       expect(addFlash).toHaveBeenCalledWith('Failed to delete 1 items.', 'error');
